+++ src/attack_surface_mapper/reporting/__init__.py
@@ -1,3 +1,4 @@
 from .generator import ReportGenerator, ReportPaths, ReportStats
+from .review_matrix import build_review_rows, review_bucket_for_finding, write_review_matrix
 
-__all__ = ['ReportGenerator', 'ReportPaths', 'ReportStats']
+__all__ = ['ReportGenerator', 'ReportPaths', 'ReportStats', 'build_review_rows', 'review_bucket_for_finding', 'write_review_matrix']